Patch by icylake: http://drupal.org/files/issues/views_datasource_alpha3_003.patch I love this module so much that I've taken the time to improve the formatting and commenting of it according to the recommendations of the Coder and Coder Tough Love modules. This patch just focuses on spacing adjustments, some comment changes, and a couple minor formatting tweaks. This is the first time I've made and submitted a patch. So, hopefully I'm doing this right !
#!
The first update was to pretty-print the JSON output when previewing it in the Views UI. The decorative and unnecessary whitespacing used in the preview will be absent when the output is retrieved outside of the Views UI.
#2
I wanted to be able to access a JSON'd view in code through the Views API without having the script terminating prematurely. The original style options wouldn't allow for this so I added "coder" variants of the original options to the style settings.

The style settings box for "JSON data document" now looks like this:
- Simple
- Simple (Coder)
- MIT Simile/Exhibit
- MIT Simile/Exhibit (Coder)

1. An improved method for placing commas between fields that takes into account empty fields that get skipped over.
2. Views UI previewing wraps the JSON preview within a tag and also formats HTML values to be displayed as regular text.
3. A much-improved views_json_get_json() function that is leaner and hopefully bug-free